fix: float to int

+++ src/discord/commands.rs
@@ -150,14 +150,14 @@ pub async fn delete_server(ctx: Context<'_>) -> Result<(), Error> {
     if new_vote == true {
         ctx.send(CreateReply::default().content(format!(
             "{:.0} left. ({:.2}%). #DELETETKGP",
-            required_votes - (yes_votes + 1.0),
+            required_votes as i64 - (yes_votes + 1),
             (yes_votes as f32 + 1.0) / required_votes * 100.0
         )))
         .await?;
     } else {
         ctx.send(CreateReply::default().content(format!(
             "{:.0} left. ({:.2}%) #DELETETKGP",
-            required_votes - (yes_votes - 1.0),
+            required_votes as i64 - (yes_votes - 1),
             (yes_votes as f32 - 1.0) / required_votes * 100.0
         )))
         .await?;
